fix: use tcp as protocol in nslookup Consul definition (#839)
Since nslookup already uses https, using http as protocol definition for Envoy results in a wrong parsing of requests that causes SSL errors on any service that tries to call nslookup via its Consul sidecar.

By using "tcp" the traffic is forwarded as-is to the service, removing the parsing errors.

Refs: GB-830